fix: display.tsが表示されないバグを修正
