+++ pages/timer/index.tsx
@@ -1,6 +1,5 @@
 import React, { VFC } from "react";
 import { Layout } from "../../components/Layout";
-import ShowDamedesu from "./script/display";
 
 import styles from "/styles/timer.module.css";
 
@@ -96,9 +95,9 @@ class RealTimeClock extends React.Component<ClockPropsType, ClockStateType> {
 }
 
 const Timer: VFC = () => {
-  ShowDamedesu();
   return (
     <Layout title="" description="">
+      {ShowDamedesu()}
       <div className={styles.backGround}>
         <div>
           <p className={styles.title}>†限界タイマー†</p>
@@ -110,3 +109,38 @@ const Timer: VFC = () => {
 };
 
 export default Timer;
+
+const ShowDamedesu = () => {
+  console.log("＿人人人人人人人人人人＿");
+  console.log("＞ 進捗ダメです！！ ＜");
+  console.log("￣Y^Y^Y^Y^Y^Y^Y^Y^Y￣");
+  console.log(
+    "Please decode! → VHN1a3VDVEZ7aHR0cHM6Ly9zZWNoYWNrMzY1Lm5pY3QuZ28uanB9"
+  );
+  console.log("                        _");
+  console.log("                      _:~~");
+  console.log("                    .::~:~.");
+  console.log("                  .:~~:~~:~");
+  console.log("                .~:~:~~::~:_");
+  console.log("              .~~:~:~:~:~~:~");
+  console.log("            .:~~:~:~~:~~:~~:~           `.");
+  console.log("          .~:~~:~~:~:~:~::~`           .uM;");
+  console.log("        . ~:~:~:~:~~~:~:~`           .JMMMN");
+  console.log("      .Mb  :~~:~~:~:~:~`           .JMMMMMM]");
+  console.log("    .MMMMc ~:~~:~~::~` ._:~__-... .MMMMMMMMM.");
+  console.log("  .MMMMMMN  ::~::~~` ._:~~:~:~:~:..MMMMMMMMMb");
+  console.log(" MMMMMMMMMb _~~:~`` _:~~:~~:~~~:~~ (MMMMMMMMMc");
+  console.log(" dMMMMMMMMM, ~:~  _:~:~:~:~~:~:~:~_ MMMMMMMMM3");
+  console.log(' .MMMMMMMMMb    .:~~:~:~~:~:~:~~:~~ ,MMMMMM"');
+  console.log('  (MMMMMMMM@`    ``~~~~:~~:~:~:~~:~_ dMMM"');
+  console.log("   HMMMMMB!                ``~~~:::~..MD");
+  console.log("   ,MMM#/           ._::~:_-...");
+  console.log("    d#^            _:~~~:~~~~:~~~~:~");
+  console.log("                  ~:~~:~~:~:~~:~:~");
+  console.log("                   ~:~:~:~:~:~:~");
+  console.log("                   _~:~~~:~~:~");
+  console.log("                    :~:~:~:~`");
+  console.log("                    _~~:~~`");
+  console.log("                     ~:~`");
+  console.log("                      `");
+};
