--- pages/timer/script/display.ts
@@ -1,36 +1,0 @@
-const ShowDamedesu = () => {
-  console.log("＿人人人人人人人人人人＿");
-  console.log("＞　進捗ダメです！！　＜");
-  console.log("￣Y^Y^Y^Y^Y^Y^Y^Y^Y￣");
-  console.log(
-    "Please decode! → VHN1a3VDVEZ7aHR0cHM6Ly9zZWNoYWNrMzY1Lm5pY3QuZ28uanB9"
-  );
-  console.log("                        _");
-  console.log("                      _:~~");
-  console.log("                    .::~:~.");
-  console.log("                  .:~~:~~:~");
-  console.log("                .~:~:~~::~:_");
-  console.log("              .~~:~:~:~:~~:~");
-  console.log("            .:~~:~:~~:~~:~~:~           `.");
-  console.log("          .~:~~:~~:~:~:~::~`           .uM;");
-  console.log("        . ~:~:~:~:~~~:~:~`           .JMMMN");
-  console.log("      .Mb  :~~:~~:~:~:~`           .JMMMMMM]");
-  console.log("    .MMMMc ~:~~:~~::~` ._:~__-... .MMMMMMMMM.");
-  console.log("  .MMMMMMN  ::~::~~` ._:~~:~:~:~:..MMMMMMMMMb");
-  console.log(" MMMMMMMMMb _~~:~`` _:~~:~~:~~~:~~ (MMMMMMMMMc");
-  console.log(" dMMMMMMMMM, ~:~  _:~:~:~:~~:~:~:~_ MMMMMMMMM3");
-  console.log(' .MMMMMMMMMb    .:~~:~:~~:~:~:~~:~~ ,MMMMMM"');
-  console.log('  (MMMMMMMM@`    ``~~~~:~~:~:~:~~:~_ dMMM"');
-  console.log("   HMMMMMB!                ``~~~:::~..MD");
-  console.log("   ,MMM#/           ._::~:_-...");
-  console.log("    d#^            _:~~~:~~~~:~~~~:~");
-  console.log("                  ~:~~:~~:~:~~:~:~");
-  console.log("                   ~:~:~:~:~:~:~");
-  console.log("                   _~:~~~:~~:~");
-  console.log("                    :~:~:~:~`");
-  console.log("                    _~~:~~`");
-  console.log("                     ~:~`");
-  console.log("                      `");
-};
-
-export default ShowDamedesu;
